+++ noj-core/src/domains/identity/routes/admin-users.ts
@@ -10,7 +10,7 @@ import {
   resolveUserId,
   unbanUser,
 } from "../services/users.ts";
-import { updateUserRoles } from "../../../services/admin-roles.ts";
+import { updateUserRoles } from "../services/admin-roles.ts";
 
 /**
  * 管理端用户管理路由（挂载前缀 /api/v1/admin，见 admin/index.ts）。
